+++ Assets/Examples/Top Down Shooter/Scripts/PlayerBody.cs
@@ -18,7 +18,7 @@ namespace PurrNet.Examples.TopDownShooter
 
             int index = 0;
             if (owner.Value.id != 0)
-                index = (int)(owner.Value.id % bodies.Count);
+                index = (int)(owner.Value.id % (ulong)bodies.Count);
 
             for (int i = 0; i < bodies.Count; i++)
             {
