+++ Assets/PurrNet/Runtime/CoreModules/Players/PlayersManager.cs
@@ -78,7 +78,7 @@ namespace PurrNet.Modules
 
         private readonly Dictionary<string, PlayerID> _cookieToPlayerId = new Dictionary<string, PlayerID>();
         private readonly Dictionary<PlayerID, string> _playerIdToCookie = new Dictionary<PlayerID, string>();
-        private ushort _playerIdCounter;
+        private ulong _playerIdCounter;
 
         private readonly Dictionary<Connection, PlayerID>
             _connectionToPlayerId = new Dictionary<Connection, PlayerID>();
